requirements.txt
# Core Dependencies
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.20
openai==1.10.0

# Vector Store & Embeddings
faiss-cpu==1.7.4
numpy==1.24.3

# Document Processing
pypdf==4.0.1
python-docx==1.1.0
unstructured==0.12.0

# Guardrails & Validation
guardrails-ai==0.4.5
pydantic==2.5.3

# API Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Environment & Configuration
python-dotenv==1.0.0

# Logging & Monitoring
structlog==24.1.0
colorama==0.4.6

# Conversation Memory
redis==5.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development Tools
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Additional Guardrails Validators
guardrails[api]==0.4.5
